feat(OpPop): impl&test vm and fix REPL
Super interesting. Because we don't actually set the value to `nil` when
we pop from the stack, the object remains in the stack. So when we
implement `LastPopped()`, we just return `vm.stack[vm.sp]` since the top
of the stack is always `vm.sp-1`, that means `vm.sp` is always the last
one popped.